♻️ Replaced `HeadingSize`, `HeadingLevel`
