+++ packages/react/src/components/Typography/Heading/Heading.stories.draft.mdx
@@ -1,5 +1,6 @@
 import { Meta, Canvas, Story } from '@storybook/addon-docs';
 import { Heading, HeadingSize, HeadingLevel } from '.';
+import { HeadingSize, HeadingLevel } from '../../../types/enums';
 import {
   ComponentUsage,
   TokensTable,
