+++ packages/react/src/components/Typography/Heading/index.ts
@@ -1,1 +1,1 @@
-export { default as Heading, HeadingSize, HeadingLevel } from './Heading';
+export { default as Heading } from './Heading';
