+++ packages/react/src/types/enums.ts
@@ -83,3 +83,20 @@ export enum RadioGroupVariant {
   Vertical = 'vertical',
   Horizontal = 'horizontal',
 }
+
+export enum HeadingSize {
+  XLarge = 'xlarge',
+  Large = 'large',
+  Medium = 'medium',
+  Small = 'small',
+  XSmall = 'xsmall',
+}
+
+export enum HeadingLevel {
+  H1 = 1,
+  H2 = 2,
+  H3 = 3,
+  H4 = 4,
+  H5 = 5,
+  H6 = 6,
+}
